Fix resize breaking input dtypes #442

+++ imgaug/augmenters/size.py
@@ -591,7 +591,7 @@ class Resize(meta.Augmenter):
         if not isinstance(images, list):
             all_same_size = (len(set([image.shape for image in result])) == 1)
             if all_same_size:
-                result = np.array(result, dtype=np.uint8)
+                result = np.array(result, dtype=images.dtype)
 
         return result
